Migrate string_utils (#60)
The built-in conversion has some problems, mainly that it recognizes
pascal case as camel case, while `lowerCamelCase()` is the actual camel
case. Plus, the actual implementation of the "camel case" doesn't work.
Therefore, the package `change_case` is introduced to solve this. For
backwards compatibility, the naming will be preserved for now.